Add Temperature Images for Windows & Linux ...
On MacOS you can set both an Icon and a Title for the Menu Bar.  Which allows you to display both the current weather condition icon ( e.g. Rainy ) and the current Temperature via Text.

Windows & Linux only allow you to set the Menu Bar Icon, which means if we want our app to show the temperature as the icon, there needs to be a temperature icon for every possible temperature :(

So I cracked open Photoshop and used their Dataset Automation process.  Then I used TinyPNG to make all of these 602 images fit within a folder that is only 384 KB. Woot Woot.